+++ src/components/primitives/color/component/index.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { isNullOrUndefined } from "@bodynarf/utils";
 
 import "./style.scss";
@@ -7,7 +9,7 @@ import ColorPickerWithoutLabel from "../components/withoutLabel";
 import ColorPickerWithLabel from "../components/withLabel";
 
 /** Color picker component */
-const ColorPicker = (props: ColorPickerProps): JSX.Element =>
+const ColorPicker: FC<ColorPickerProps> = (props) =>
     isNullOrUndefined(props.label)
         ? <ColorPickerWithoutLabel {...props} />
         : <ColorPickerWithLabel {...props} />