+++ src/components/primitives/multiline/component/index.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { isNullOrUndefined } from "@bodynarf/utils";
 
 import { MultilineProps } from "..";
@@ -5,7 +7,7 @@ import MultilineWithoutLabel from "../components/withoutLabel";
 import MultilineWithLabel from "../components/withLabel";
 
 /** Multiline textual input component */
-const Multiline = (props: MultilineProps): JSX.Element => {
+const Multiline: FC<MultilineProps> = (props) => {
     return isNullOrUndefined(props.label)
         ? <MultilineWithoutLabel {...props} />
         : <MultilineWithLabel {...props} />;