+++ src/components/primitives/number/component/index.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { isNullOrUndefined } from "@bodynarf/utils";
 
 import { NumberProps } from "..";
@@ -5,7 +7,7 @@ import NumberWithLabel from "../components/withLabel";
 import NumberWithoutLabel from "../components/withoutLabel";
 
 /** Number input component */
-const Number = (props: NumberProps): JSX.Element => {
+const Number: FC<NumberProps> = (props) => {
     return isNullOrUndefined(props.label)
         ? <NumberWithoutLabel {...props} />
         : <NumberWithLabel {...props} />;