+++ src/components/primitives/password/component/index.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { isNullOrUndefined } from "@bodynarf/utils";
 
 import "./style.scss";
@@ -7,7 +9,7 @@ import PasswordWithLabel from "../components/withLabel";
 import PasswordWithoutLabel from "../components/withoutLabel";
 
 /** Password input component */
-const Password = (props: PasswordProps): JSX.Element => {
+const Password: FC<PasswordProps> = (props) => {
     return isNullOrUndefined(props.label)
         ? <PasswordWithoutLabel {...props} />
         : <PasswordWithLabel {...props} />;