+++ src/components/primitives/password/components/withoutLabel/index.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useCallback, useState } from "react";
+import { ChangeEvent, FC, useCallback, useState } from "react";
 
 import { emptyFn, generateGuid, getClassName, isNullOrUndefined } from "@bodynarf/utils";
 
@@ -9,7 +9,7 @@ import InternalHint from "@bbr/internalComponent/hint";
 
 import { PasswordProps } from "../..";
 
-const PasswordWithoutLabel = ({
+const PasswordWithoutLabel: FC<PasswordProps> = ({
     defaultValue,
     onValueChange = emptyFn, disabled, validationState,
     name = generateGuid(),
@@ -20,7 +20,7 @@ const PasswordWithoutLabel = ({
 
     className, title, data,
     hint,
-}: PasswordProps): JSX.Element => {
+}) => {
     const onChange = useCallback(
         (event: ChangeEvent<HTMLInputElement>) => onValueChange(event.target.value),
         [onValueChange]