+++ src/components/primitives/text/component/index.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { isNullOrUndefined } from "@bodynarf/utils";
 
 import { TextProps } from "..";
@@ -5,7 +7,7 @@ import TextWithLabel from "../components/withLabel";
 import TextWithoutLabel from "../components/withoutLabel";
 
 /** Textual input component */
-const Text = (props: TextProps): JSX.Element => {
+const Text: FC<TextProps> = (props) => {
     return isNullOrUndefined(props.label)
         ? <TextWithoutLabel {...props} />
         : <TextWithLabel {...props} />;