+++ changelog.md
@@ -45,6 +45,7 @@ All changes will be published here in reverse chronological order
 
 - **All input controls**
   - Updated the `onValueChange` callback signature to support `nullable` values, depending on the component.
+  - Updated the `ValidationState`: `messages` property has been made optional.
 
 - **Checkbox**
   - Updated requirements for members of the `label` prop.